requirements.txt
fastapi
uvicorn
pymilvus
weaviate-client
psycopg2-binary
